collision aek le bord de la map
ca ne bug plus quand on est trop proche du bors

+++ src/define.h
@@ -9,8 +9,8 @@
 #define NB_BLOCS_X 11*(float)((float)64/BLOCS_SIZE)
 #define NB_BLOCS_Y 9*(float)((float)64/BLOCS_SIZE)
 
-#define DIST_FROM_UP (float)((float)(NB_BLOCS_Y-1)/2)+1
-#define DIST_FROM_LEFT (float)((float)(NB_BLOCS_X-1)/2)+1
+#define DIST_FROM_UP (int)((float)(NB_BLOCS_Y-1)/2)+1
+#define DIST_FROM_LEFT (int)((float)(NB_BLOCS_X-1)/2)+1
 
 //jai ajouter une coupe de (float) pour etre sur
